Se craron los objetos de valor y validaciones para la entidad Bebida
