+++ src/libs/validations/index.ts
@@ -5,6 +5,7 @@ export * from './es-guarnicion-vlaida.validations';
 export * from './es-nombre.validations';
 export * from './es-precio-valido.validations';
 export * from './es-string-alfabetico.validations';
+export * from './es-tamaño-valido.validations';
 export * from './es-uuid.validations';
 export * from './es-vehiculo.validations';
 export * from './longitud-maxima.validations';
