+++ src/subdomains/servicio-al-cliente/contexts/entrega-a-domicilio/domain/value-objects/pedido/bebida/bebida-id/index.ts
@@ -1,0 +1,2 @@
+
+export * from './bebida-id.value-object';
