--- src/subdomains/servicio-al-cliente/contexts/entrega-a-domicilio/domain/value-objects/pedido/bebida/cliente-id/index.ts
@@ -1,2 +1,0 @@
-
-export * from './cliente-id.value-object';
