--- src/subdomains/servicio-al-cliente/contexts/entrega-a-domicilio/domain/value-objects/pedido/bebida/direccion/direccion.value-object.ts
@@ -1,39 +1,0 @@
-import { EsDireccion } from '../../../../../../../../../libs/validations';
-import { ValueObjectBase } from '../../../../../../../../../libs/sofka/bases';
-import { IErrorValueObject } from '../../../../../../../../../libs/sofka/interface';
-
-
-export class DireccionValueObject extends ValueObjectBase<string> {
-
-    constructor(value: string) {
-        super(value);
-    }
-    
-    /**
-     *Funsión que realiza todas las validaciones
-     *
-     * @memberof DireccionValueObject
-     */
-    validateData(): void {
-        this.validarEstructura();
-    }
-
-    /**
-     *Funsión que valida la estructura de una dirección
-     *
-     * @private
-     * @memberof DireccionValueObject
-     */
-    private validarEstructura(): void {
-        if(EsDireccion(this.value) === false) {
-
-            const error: IErrorValueObject = {
-                field: 'Direccion',
-                message: 'El valor no es una dirección válida'
-            };
-
-            this.setError(error);
-        }
-    }
-    
-}
