+++ src/subdomains/servicio-al-cliente/contexts/entrega-a-domicilio/domain/value-objects/pedido/bebida/index.ts
@@ -1,3 +1,3 @@
-export * from './cliente-id';
-export * from './direccion';
+export * from './bebida-id';
+export * from './tamanio';
 export * from './nombre';
