+++ src/subdomains/servicio-al-cliente/contexts/entrega-a-domicilio/domain/value-objects/pedido/bebida/tamanio/index.ts
@@ -1,0 +1,2 @@
+
+export * from './tamanio.value-object';
